+++ src/main/java/de/codengine/krabat/anims/Mlynk2.java
@@ -308,7 +308,7 @@ public class Mlynk2 extends Mainanim {
         int scale = getScale((int) yps);
 
         // Zooming - Faktor beruecksichtigen in y-Richtung
-        float vert_dist = CVERT_DIST - scale / SLOWY;
+        float vert_dist = CVERT_DIST - (float) scale / SLOWY;
         if (vert_dist < 1) {
             vert_dist = 1;
             // hier kann noch eine Entscheidungsroutine hin, die je nach Animationsphase
