+++ src/main/java/de/codengine/krabat/locations2/Kulow2.java
@@ -257,7 +257,7 @@ public class Kulow2 extends Mainloc {
 
         // Parallaxer fuer Saeule, muss immer Hintergrund loeschen ?????
         float xtf = mainFrame.scrollx;
-        xtf = 1180 - mainFrame.scrollx * 3 / 8;
+        xtf = 1180 - (float) (mainFrame.scrollx * 3) / 8;
         int xt = (int) xtf;
         g.setClip(xt - 2, 208, xt + 90, 479);
         g.drawImage(backright, 640, 0);
